Extract the code launching the runner out of the MigrationTest class to be able to run it using mvn exec or java -jar
